Feature/fe/etc 계정탈퇴alert

See merge request sw_track/class_03/web_project_2/team17/bobhub!120

+++ FE/src/pages/MyPage/components/DeleteUser.tsx
@@ -14,11 +14,15 @@ const DeleteUser = () => {
 
   const handleUserDelete = (e: React.MouseEvent<HTMLElement>) => {
     e.preventDefault();
-    try {
-      deleteUserAPI();
-      alert('계정 탈퇴되었습니다.');
-    } catch (err) {
-      console.error(err);
+    if (confirm('정말 계정을 삭제하시겠습니까?') === false) {
+      alert('계정 삭제를 취소하셨습니다.');
+    } else {
+      try {
+        deleteUserAPI();
+        alert('계정 탈퇴되었습니다.');
+      } catch (err) {
+        console.error(err);
+      }
     }
   };
